widen unit energy type

+++ definitions.h
@@ -8,7 +8,7 @@ typedef uint8_t unit_health;
 typedef uint8_t unit_type;
 typedef uint8_t grid_index;
 typedef uint16_t queue_index;
-typedef uint8_t unit_energy;
+typedef uint16_t unit_energy;
 
 #define queue_capacity 0xffff
 #define units_capacity (unit_index)'\xff'
